Update UI #2 - Update material2 to material3 - Add FavoriteTab - Change ListScreen to ListTab
