+++ shared/src/commonMain/kotlin/presentation/tabs/MainScreen.kt
@@ -1,0 +1,12 @@
+package presentation.tabs
+
+import androidx.compose.runtime.Composable
+import cafe.adriel.voyager.core.screen.Screen
+
+internal object MainScreen : Screen {
+
+    @Composable
+    override fun Content() {
+        // TODO
+    }
+}
