+++ shared/src/commonMain/kotlin/presentation/tabs/favorite/FavoriteTab.kt
@@ -1,0 +1,42 @@
+package presentation.tabs.favorite
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Favorite
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.rememberVectorPainter
+import cafe.adriel.voyager.navigator.tab.Tab
+import cafe.adriel.voyager.navigator.tab.TabOptions
+
+internal object FavoriteTab : Tab {
+
+    override val options: TabOptions
+        @Composable
+        get() {
+            val title = "Favorite"
+            val icon = rememberVectorPainter(Icons.Default.Favorite)
+
+            return remember {
+                TabOptions(
+                    index = 0u,
+                    title = title,
+                    icon = icon
+                )
+            }
+        }
+
+
+    @Composable
+    override fun Content() {
+        Box(
+            modifier = Modifier
+                .fillMaxSize(),
+        ) {
+            Text(text = "Favorite")
+        }
+    }
+}
