+++ shared/src/commonMain/kotlin/presentation/tabs/list/ListContract.kt
@@ -1,16 +1,17 @@
-package presentation.detail
+package presentation.tabs.list
 
+import domain.model.CountryModel
 import presentation.base.BaseViewAction
 import presentation.base.BaseViewEffect
 import presentation.base.BaseViewState
 
-class DetailContract {
+class ListContract {
 
     data class State(
         val isLoading: Boolean = false,
         val isError: Boolean = false,
         val errorMessage: String = "",
-        val detail: String = "",
+        val list: List<CountryModel> = emptyList(),
     ) : BaseViewState
 
     sealed interface Action : BaseViewAction {
@@ -19,8 +20,11 @@ class DetailContract {
 
         data class Error(val errorMessage: String) : Action
 
-        data class Success(val detail: String) : Action
+        data class Success(val list: List<CountryModel>) : Action
     }
 
-    sealed interface Effect : BaseViewEffect
+    sealed interface Effect : BaseViewEffect {
+
+        data class OpenDetailScreen(val id: String) : Effect
+    }
 }
