+++ shared/src/commonMain/kotlin/presentation/ui/Theme.kt
@@ -1,6 +1,6 @@
 package presentation.ui
 
-import androidx.compose.material.MaterialTheme
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 
 @Composable
